Uploading first pass of Controller files.
There's some issue with my git config and I may have to reset a bunch of credential. Before I do that, I'm just going to ensure that my work is saved to a branch.

Code for the first pass of the Controller files is contained in this checking. Note, Ansh already hit Stores so that was done.